Remove labextension from setup.py
