+++ bindings/python/pydeck/setup.py
@@ -9,7 +9,6 @@ from setuptools.command.develop import develop
 
 import atexit
 from distutils import log
-import glob
 import json
 import os
 from shutil import copy
@@ -83,7 +82,6 @@ class FrontendBuild(Command):
     target_files = [
         os.path.join(here, "pydeck", "nbextension", "static", "index.js"),
         os.path.join(here, "pydeck", "nbextension", "static", "index.js.map"),
-        os.path.join(here, "pydeck", "labextension/"),
     ]
 
     user_options = []
@@ -115,8 +113,6 @@ class FrontendBuild(Command):
            Overwrites destination files"""
         js_dist_dir = os.path.join(widget_dir, "dist")
         nbextension_folder = os.path.join(here, "pydeck", "nbextension", "static")
-        labextension_folder = os.path.join(here, "pydeck", "labextension")
-        labextension_tgz = glob.glob(os.path.join(widget_dir, "deck.gl-jupyter-widget-*.tgz"))[0]
         js_files = [
             {
                 "source": os.path.join(js_dist_dir, "index.js"),
@@ -125,10 +121,6 @@ class FrontendBuild(Command):
             {
                 "source": os.path.join(js_dist_dir, "index.js.map"),
                 "destination": nbextension_folder,
-            },
-            {
-                "source": labextension_tgz,
-                "destination": labextension_folder
             },
         ]
         for js_file in js_files:
@@ -157,13 +149,6 @@ class FrontendBuild(Command):
                 env=env,
             )
             log.info("Creating lab extension .tgz with `npm pack`.")
-            check_call(
-                ["npm", "run", "build:labextension"],
-                cwd=widget_dir,
-                stdout=sys.stdout,
-                stderr=sys.stderr,
-                env=env,
-            )
         else:
             log.info("Installing build dependencies with `npm run build`.")
             check_call(
@@ -281,8 +266,6 @@ if __name__ == "__main__":
                     "pydeck/io/templates/requirejs_dependencies.json",
                     "pydeck/nbextension/static/index.js.map",
                 ]),
-            ('share/jupyter/lab/extensions', ['pydeck/labextension/*']),
-
             ("etc/jupyter/nbconfig/notebook.d", ["pydeck.json"]),
         ],
         zip_safe=False,
